+++ source/ReadAlign_alignBAM.cpp
@@ -96,6 +96,14 @@ int ReadAlign::alignBAM(Transcript const &trOut, uint nTrOut, uint iTrOut, uint 
         calcCIGAR(trOut, nMates, iExMate, leftMate);
     };
 
+
+    int32 gxgnGene = -1; //gene to output in GX/GN
+    if (P.outSAMattrPresent.GX || P.outSAMattrPresent.GN) {
+        auto annFeat = readAnnot.annotFeatures[P.pSolo.samAttrFeature];
+        if (annFeat.fSet.size()==1 && annFeat.fAlign[iTrOut]>0)
+            gxgnGene = annFeat.fAlign[iTrOut];
+    };
+
     for (uint imate=0;imate < (alignType<0 ? nMates:P.readNmates);imate++) { //not readNends: this is about alignments
 
         uint iEx1=0;
@@ -384,40 +392,13 @@ int ReadAlign::alignBAM(Transcript const &trOut, uint nTrOut, uint iTrOut, uint 
                         break;
 
                     case ATTR_GX:
-                        //TODO: for multimappers: output GX,GN with M, prefix and all aigns for each trOut
-                        // for non-genic, output -
-                        // for geneFull: need trOut.alignGeneFull to output GX only for genic aligns
-                        if ( P.quant.gene.yes ) {
-                            if (readAnnot.geneConcordant.size()==1 && trOut.alignGenes.size()==1) 
-                                //only output if read maps to a single gene, and this trOut maps to one gene - it will be the same gene
-                                //if a read maps to multiple genomic loci, but only one gene, only one trOut will have trOut.alignGenes
-                                attrN+=bamAttrArrayWrite(chunkTr->geID[*trOut.alignGenes.begin()],"GX",attrOutArray+attrN);
-                        } else if ( P.quant.geneFull.yes ) {
-                            if ( readAnnot.geneFull.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geID[*readAnnot.geneFull.begin()],"GX",attrOutArray+attrN);
-                        } else if ( P.quant.geneFull_ExonOverIntron.yes ) {
-                            if ( readAnnot.geneFull_ExonOverIntron.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geID[*readAnnot.geneFull_ExonOverIntron.begin()],"GX",attrOutArray+attrN);                                
-                        } else if ( P.quant.geneFull_Ex50pAS.yes ) {
-                            if ( readAnnot.geneFull_Ex50pAS.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geID[*readAnnot.geneFull_Ex50pAS.begin()],"GX",attrOutArray+attrN);                            
-                        };
+                        if (gxgnGene > 0)
+                            attrN+=bamAttrArrayWrite(chunkTr->geID[gxgnGene],"GX",attrOutArray+attrN);
                         break;
                             
                     case ATTR_GN:
-                        if ( P.quant.gene.yes ) {                        
-                            if (readAnnot.geneConcordant.size()==1 && trOut.alignGenes.size()==1) //only output if read maps to a single gene, and trOut maps to one gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geName[*trOut.alignGenes.begin()],"GN",attrOutArray+attrN);
-                        } else if ( P.quant.geneFull.yes ) {
-                            if ( readAnnot.geneFull.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geName[*readAnnot.geneFull.begin()],"GN",attrOutArray+attrN);
-                        } else if ( P.quant.geneFull_ExonOverIntron.yes ) {
-                            if ( readAnnot.geneFull_ExonOverIntron.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geName[*readAnnot.geneFull_ExonOverIntron.begin()],"GN",attrOutArray+attrN);                                
-                        } else if ( P.quant.geneFull_Ex50pAS.yes ) {
-                            if ( readAnnot.geneFull_Ex50pAS.size()==1 ) //only output if read maps to a single gene
-                                attrN+=bamAttrArrayWrite(chunkTr->geName[*readAnnot.geneFull_Ex50pAS.begin()],"GN",attrOutArray+attrN);                            
-                        };                           
+                        if (gxgnGene > 0)
+                            attrN+=bamAttrArrayWrite(chunkTr->geName[gxgnGene],"GN",attrOutArray+attrN);
                         break;                        
                         
                     case ATTR_sM: