+++ source/Transcriptome_geneFullAlignOverlap_ExonOverIntron.cpp
@@ -11,6 +11,7 @@ void Transcriptome::geneFullAlignOverlap_ExonOverIntron(uint nA, Transcript **aA
 
     //calculate overlap with introns
     annFeat.fSet={};
+    annFeat.fAlign = {};    
     annFeat.fAlign.resize(nA,-1);
 
     for (uint32 iA=0; iA<nA; iA++) {//only includes aligns that are entirely inside genes (?)